+++ src/camera/Camera.ts
@@ -9,6 +9,7 @@ import Fade from '../effects/Fade';
 import Shake from '../effects/Shake';
 import PanTo from '../effects/PanTo';
 import ZoomTo from '../effects/ZoomTo';
+import Rotate from '../effects/Rotate';
 import Effect from '../effects/Effect';
 
 /**
@@ -121,6 +122,19 @@ export default class Camera {
   }
 
   /**
+   * Rotates to a specified angle.
+   * 
+   * @param {number} angle The angle to rotate to, from 0 to 360 with 0 being the default state and 360 being all the way around back to the default state.
+   * @param {number} duration The amount of time, in milliseconds, that the effect should take.
+   * @param {Function} [easing=easeLinear] The easing function that should be used.
+   */
+  rotate(angle: number, duration: number, easing: Function = easeLinear) {
+    const rotate: Rotate = new Rotate(this._container, angle, duration, easing);
+
+    this._addToTicker(rotate);
+  }
+
+  /**
    * Adds an effect to the PIXI Ticker if it's being used and removes it when necessary.
    * 
    * @private
@@ -161,7 +175,7 @@ export default class Camera {
    */
   private _setupFilter() {
     this._filter = new this._options.sprite(this._options.texture.WHITE);
-    
+
     this._filter.width = this._container.width;
     this._filter.height = this._container.height;
     this._filter.alpha = 0;